Fix for issues saving custom rewards or reqs
I think there's probably a way this could be fixed upstream, but I have
to admit I'm not familiar with the Bukkit chat session API. It seems
like the Quest itself loads this data in as a Map, but the chat session
wants to set it as a List for a new Quest.

My solution is to just handle both, but this feels like a hack.